pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "regulation-compliance-agent"
version = "0.1.0"
description = "資安合規代理人系統 - 多代理人法規蒐集與解析平台"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "YANG JUI HSIU", email = "yang06140425@gmail.com"}
]
keywords = ["ai", "agents", "compliance", "regulation", "azure", "autogen"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pyautogen>=0.2.0",
    "openai>=1.12.0",
    "gradio>=4.0.0",
    "httpx>=0.26.0",
    "beautifulsoup4>=4.12.0",
    "pydantic>=2.5.0",
    "chromadb>=0.4.20",
    "sqlalchemy>=2.0.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "mypy>=1.8.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
]
crawler = [
    "scrapy>=2.11.0",
    "playwright>=1.40.0",
]

[project.scripts]
compliance-agent = "app:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "config*"]

[tool.black]
line-length = 88
target-version = ['py310', 'py311', 'py312']

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
